apps/pptx_processor/requirements.txt
# Core Dataloop SDK
dtlpy>=1.0.0

# Data Processing
pandas>=2.0.0

# PPTX Processing
python-pptx>=0.6.21

# Text Processing
unstructured>=0.10.0
nltk>=3.8.0
langchain-text-splitters>=0.0.1
